Move copy scores from text dataset to copy_generator.
