+++ onmt/inputters/text_dataset.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
 
-import torch
-
 from onmt.inputters.dataset_base import DatasetBase
 
 
@@ -25,35 +23,6 @@ class TextDataset(DatasetBase):
         if hasattr(ex, "tgt"):
             return len(ex.src), len(ex.tgt)
         return len(ex.src)
-
-    @staticmethod
-    def collapse_copy_scores(scores, batch, tgt_vocab, src_vocabs,
-                             batch_dim=1, batch_offset=None):
-        """
-        Given scores from an expanded dictionary
-        corresponeding to a batch, sums together copies,
-        with a dictionary word when it is ambiguous.
-        """
-        offset = len(tgt_vocab)
-        for b in range(scores.size(batch_dim)):
-            blank = []
-            fill = []
-            batch_id = batch_offset[b] if batch_offset is not None else b
-            index = batch.indices.data[batch_id]
-            src_vocab = src_vocabs[index]
-            for i in range(1, len(src_vocab)):
-                sw = src_vocab.itos[i]
-                ti = tgt_vocab.stoi[sw]
-                if ti != 0:
-                    blank.append(offset + i)
-                    fill.append(ti)
-            if blank:
-                blank = torch.Tensor(blank).type_as(batch.indices.data)
-                fill = torch.Tensor(fill).type_as(batch.indices.data)
-                score = scores[:, b] if batch_dim == 1 else scores[b]
-                score.index_add_(1, fill, score.index_select(1, blank))
-                score.index_fill_(1, blank, 1e-10)
-        return scores
 
     @classmethod
     def make_examples(cls, sequences, side):
